// src/pages/main-page/model/types.ts
enum FILTER_TYPE {
	NONE = 'NONE',
	NOUN = 'имя существительное',
	ADJF = 'имя прилагательное (полное)',
	ADJS = 'имя прилагательное (краткое)',
	COMP = 'компаратив',
	VERB = 'глагол (личная форма)',
	INFN = 'глагол (инфинитив)',
	PRTF = 'причастие (полное)',
	PRTS = 'причастие (краткое)',
	GRND = 'деепричастие',
	NUMR = 'числительное',
	ADVB = 'наречие',
	NPRO = 'местоимение-существительное',
	PRED = 'предикатив',
	PREP = 'предлог',
	CONJ = 'союз',
	PRCL = 'частица',
	INTJ = 'междометие',
}

export { FILTER_TYPE }